Stop blender from rendering when cancel is pressed. Also, prevent Animated Title screen from jumping focus back to listview after thumbnail preview is generated

+++ src/windows/views/blender_listview.py
@@ -33,6 +33,7 @@ import functools
 import shlex
 import json
 from time import sleep
+import signal
 
 # Try to get the security-patched XML functions from defusedxml
 try:
@@ -69,10 +70,14 @@ class BlenderListView(QListView):
         self.selected = selected
         self.deselected = deselected
 
+        if not selected or not selected.isValid():
+            return
+        if getattr(self, "_last_selected_index", None) == selected:
+            return
+
         # Get translation object
         _ = self.app._tr
 
-        self.win.clear_effect_controls()
         animation = self.get_animation_details()
         self.selected_template = animation.get("service")
 
@@ -80,6 +85,9 @@ class BlenderListView(QListView):
         # but the selection is -1. So, just do nothing here.
         if not self.selected_template:
             return
+
+        self._last_selected_index = selected
+        self.win.clear_effect_controls()
 
         # Assign a new unique id for each template selected
         self.generateUniqueFolder()
@@ -266,11 +274,11 @@ class BlenderListView(QListView):
         if not os.path.exists(os.path.join(info.BLENDER_PATH, self.unique_folder_name)):
             os.mkdir(os.path.join(info.BLENDER_PATH, self.unique_folder_name))
 
-    def processing_mode(self, cursor=True):
+    def processing_mode(self, cursor=True, restore_focus=True):
         """ Disable all controls on interface """
 
-        # Store keyboard-focused widget
-        self.focus_owner = self.win.focusWidget()
+        # Store keyboard-focused widget when we plan to restore it
+        self.focus_owner = self.win.focusWidget() if restore_focus else None
 
         self.win.btnRefresh.setEnabled(False)
         self.win.sliderPreview.setEnabled(False)
@@ -624,7 +632,7 @@ Blender Path: {}
         """ Render an images sequence of the current template using Blender 2.62+ and the
         Blender Python API. """
 
-        self.processing_mode()
+        self.processing_mode(restore_focus=frame is None)
 
         # Init blender paths
         blend_file_path = os.path.join(
@@ -693,6 +701,7 @@ Blender Path: {}
         self._color_scratchpad = None
         self.selected_template = ""
         self.final_render = False
+        self._last_selected_index = None
 
         # Preview render timer
         self.preview_timer = QTimer(self)
@@ -786,18 +795,36 @@ class Worker(QObject):
             log.debug('Removing custom LD_LIBRARY_PATH from environment variables when launching Blender')
 
         self.startupinfo = None
+        self.creationflags = 0
         if sys.platform == 'win32':
             self.startupinfo = subprocess.STARTUPINFO()
             self.startupinfo.dwFlags |= subprocess.STARTF_USESHOWWINDOW
+            self.creationflags = subprocess.CREATE_NEW_PROCESS_GROUP
 
     def Cancel(self):
         """Cancel worker render"""
-        if self.process:
-            while self.process and self.process.poll() == None:
-                log.debug("Terminating Blender Process")
-                self.process.terminate()
-                sleep(0.1)
         self.canceled = True
+        if self.process and self.process.poll() is None:
+            log.debug("Terminating Blender Process")
+            try:
+                if sys.platform != "win32":
+                    os.killpg(self.process.pid, signal.SIGTERM)
+                else:
+                    self.process.terminate()
+            except Exception:
+                pass
+            for _ in range(30):
+                if not self.process or self.process.poll() is not None:
+                    break
+                sleep(0.05)
+            if self.process and self.process.poll() is None:
+                try:
+                    if sys.platform != "win32":
+                        os.killpg(self.process.pid, signal.SIGKILL)
+                    else:
+                        self.process.kill()
+                except Exception:
+                    pass
 
     def blender_version_check(self):
         # Check the version of Blender
@@ -919,6 +946,10 @@ class Worker(QObject):
 
         _ = get_app()._tr
 
+        if self.canceled:
+            self.finished.emit()
+            return
+
         if not self.version and not self.blender_version_check():
             self.finished.emit()
             return
@@ -955,10 +986,20 @@ class Worker(QObject):
             self.process = subprocess.Popen(
                 command_render, bufsize=512,
                 stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
-                startupinfo=self.startupinfo, env=self.env, cwd=info.HOME_PATH
+                startupinfo=self.startupinfo,
+                creationflags=self.creationflags,
+                start_new_session=(sys.platform != "win32"),
+                env=self.env,
+                cwd=info.HOME_PATH,
             )
             # Signal UI that background task is running
             self.start_processing.emit()
+
+            if self.canceled:
+                self.Cancel()
+                self.end_processing.emit()
+                self.finished.emit()
+                return
 
         except subprocess.SubprocessError as ex:
             # Error running command.  Most likely the blender executable path in
